Moves postgres function tests from unit tests to integration tests
