+++ src/test/integration/resources/application.properties
@@ -14,3 +14,7 @@ idam.api.url.jwkUrl=http://localhost:5000/jwks
 idam.s2s-auth.microservice=et_cos
 idam.s2s-auth.totp_secret=AAAAAAAAAAAAAAAC
 idam.s2s-auth.url=http://localhost:8080
+
+spring.datasource.url=${DB_URL}
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
